Implement Firebase integration for user data management in Home component
