+++ you-jujube/src/Home.js
@@ -2,10 +2,13 @@ import React, { useState, useEffect } from "react";
 import VideoCard from "./components/VideoCard";
 import { getVideos } from "./services/videoService";
 import UserForm from "./components/UserForm";
+import { useAuth0 } from '@auth0/auth0-react';
+import { db, getDoc, setDoc, doc } from "./firebase-config"; 
 
 const Home = () => {
   const [videos, setVideos] = useState([]);
   const [userData, setUserData] = useState(null);
+  const { user, isAuthenticated, isLoading } = useAuth0();
 
   useEffect(() => {
     const fetchVideos = async () => {
@@ -16,8 +19,17 @@ const Home = () => {
     fetchVideos();
   }, []);
 
-  const handleFormSubmit = (data) => {
+  const handleFormSubmit = async (data) => {
     setUserData(data);
+    console.log(data);
+
+    try {
+      const userRef = doc(db, "users", user.sub);
+      await setDoc(userRef, data, { merge: true });
+      console.log("User data saved to Firebase");
+    } catch (error) {
+      console.error("Error saving user data to Firebase:", error);
+    }
   };
 
   if (!userData) {
@@ -26,6 +38,10 @@ const Home = () => {
         <UserForm onSubmit={handleFormSubmit} />
       </div>
     );
+  };
+
+  if (isLoading) {
+    return <div>Loading ...</div>;
   }
 
   return (
